BIG CHANGES
Simplifying the content model.

Initially when building this, my thought was that we should have special
parsers for each type of content (blog posts, videos, etc.), so that we
can generate special prompts to make social blurbs better.

After going down that path, however, it didn't make sense as the
application was scaling up. It makes more sense to simply have a single
type of "Content", then to use LLMs to figure out the best way to
promote the content with minimal input.

This simplifies the codebase quite a bit and makes it far easier to
customize in the future with no change to quality.