Adding notification functionality.
The way this works is -- if there is new content available to promote,
users of the app who have at least one social connection will get an
email with a link to the content they can choose to promote.

They can then click a link in the email and have it take them back to
the promotion web app where a new social blurb will be generated for
them and they can choose to one-click share the post or make edits to
it.